memok8rt-payment/src/main/java/com/sudipta/payment/serviceImpl/OrderItemServiceImplementation.java
package com.sudipta.payment.serviceImpl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.sudipta.payment.modal.OrderItem;
import com.sudipta.payment.repository.OrderItemRepository;
import com.sudipta.payment.service.OrderItemService;

@Service
public class OrderItemServiceImplementation implements OrderItemService {
	@Autowired
	private OrderItemRepository orderItemRepository;
	
	@Override
	public OrderItem createOrderItem(OrderItem orderItem) {
		
		return orderItemRepository.save(orderItem);
	}

}
